added commissioning screen link to the home screen
